+++ src/screens/LoginScreen/LoginScreen.tsx
@@ -7,10 +7,10 @@ import { Button } from 'react-native-elements';
 import { useDispatch } from 'react-redux';
 import * as Yup from 'yup';
 
-import TextField from '../components/form/TextField';
-import { NavigationData } from '../navigation/AuthNavigator';
-import { login } from '../redux/auth';
-import Colors from '../theme/Colors';
+import TextField from '../../components/form/TextField';
+import { NavigationData } from '../../navigation/AuthNavigator';
+import { login } from '../../redux/auth';
+import Colors from '../../theme/Colors';
 
 import { StyledView, StyledWideContainer } from './LoginScreen.styled';
